# openapi.yaml
openapi: 3.1.0
info:
  title: Spyglass IDX API
  description: |
    Real estate listing search API for the Spyglass IDX application.
    Proxies MLS data from Repliers API and lead capture to Follow Up Boss CRM.
  version: 0.1.0
  contact:
    name: Spyglass Realty Engineering
servers:
  - url: https://spyglass-idx.vercel.app
    description: Production
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/nlp-search:
    post:
      summary: AI-powered natural language search
      description: |
        Interprets a natural language prompt (e.g., "3 bed house in Barton Hills under 600k")
        using the Repliers NLP endpoint, optionally matches a community polygon for geographic
        filtering, and returns listings matching the interpreted search.
        
        Security: Input is validated for length (max 500 chars) and prompt injection patterns.
      operationId: nlpSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
                  maxLength: 500
                  description: Natural language real estate search query
                  example: "3 bedroom house in Barton Hills under $600k"
                nlpId:
                  type: string
                  description: Conversation ID for follow-up queries (from previous response)
      responses:
        "200":
          description: NLP search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  nlpId:
                    type: string
                    description: Conversation ID for follow-up queries
                  summary:
                    type: string
                    description: Human-readable summary of interpreted search
                  searchParams:
                    type: object
                    additionalProperties:
                      type: string
                    description: Parsed search parameters from NLP
                  listings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                  total:
                    type: integer
                  matchedCommunity:
                    type: object
                    nullable: true
                    properties:
                      name:
                        type: string
                      slug:
                        type: string
                    description: Matched community polygon info, if any
                  imageSearch:
                    type: object
                    nullable: true
                    description: Image search items from NLP, if any
        "400":
          description: Invalid input (empty prompt, too long, injection detected)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "406":
          description: Prompt not related to real estate
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: not_real_estate
                  message:
                    type: string
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Listing:
      type: object
      properties:
        id:
          type: string
          description: Same as mlsNumber
        mlsNumber:
          type: string
          example: "T1234567"
        address:
          type: object
          properties:
            street:
              type: string
              example: "123 Main St"
            city:
              type: string
              example: "Austin"
            state:
              type: string
              example: "TX"
            zip:
              type: string
              example: "78704"
            full:
              type: string
              example: "123 Main St, Austin, TX 78704"
            neighborhood:
              type: string
              nullable: true
        price:
          type: integer
          example: 450000
        bedrooms:
          type: integer
          example: 3
        bathrooms:
          type: integer
          example: 2
        sqft:
          type: integer
          example: 1800
        lotSize:
          type: number
          nullable: true
        yearBuilt:
          type: integer
          nullable: true
        propertyType:
          type: string
          enum: [Single Family, Condo, Townhouse, Multi-Family, Land]
        status:
          type: string
          enum: [Active, Pending, Sold, Coming Soon]
        daysOnMarket:
          type: integer
        photos:
          type: array
          items:
            type: string
            format: uri
        description:
          type: string
          nullable: true
        coordinates:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
        listDate:
          type: string
          format: date
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
          nullable: true